Basic balancing using I2Cdev library with ypr
Added I2Cdev library and tuned PID constants to achieve basic balancing.
Robot can balance for a minute or so, but this depends on being placed
on a stand on power-up to determine the pitch angle setpoint (since the
returned pitch is relative to the position on power-up).

Sometimes the robot seems to slowly oscillate and eventually fall over.
Further tuning of the PID parameters (and maybe sample time) will
hopefully rectify this.

Also further tuned dead band parameters for the motors. However, the
motors still move a little independently when moving at slow speeds.