Balancing improvements - PID, balance point
Altered PID constants for faster control loop (4ms vs 5ms) and fresh batteries (1.6v AA alkaline). All gains are lower than previous iteration.

Also implemented a balance setpoint adjustment which changes the pitch setpoint to try to stop the robot wandering. If the robot is not stationary, change the setpoint to try to make it stationary.

Added a deadband to stop small motor movements when the robot is balanced.

Robot seems more stable than before and does not move around as much.