Basic forwards and backwards driving
Accept a command over serial that lasts for 5 seconds. F changes the pitch setpoint to a small positive value, and B changes it to a small negative value. Monitor the output of the PID controller and adjust the setpoint to ensure that the motors are moving, but try to prevent them moving too fast (which would make the robot fall over).

Also increased PID gain Kd to reduce oscillations from dynamic disturbance.