Improved speed control with PID controller
Added a second PID controller which measures speed and controls the pitch angle setpoint. This replaces both the previous balance setpoint code and the drive setpoint code. Now the remote command simply sets a target speed and the PID controller changes the pitch angle setpoint to achieve that speed.

Further tuning of this PID controller is required, as is some filtering of the speed value to reduce noise.