New remote control using analogue joystick
I built a remote control using an Arduino Nano, an analogue joystick and an HC-05 bluetooth module. The HC-05 seems to have a buffering mechanism which stores characters in a 230 byte buffer. It transmits when ~120ms passes without sending a byte, or when the buffer is full. To work around this, I changed the command timer on the robot to 150ms to account for both the Android app and the remote control.