Adjusted constants for soldered circuit board
Moved the robot controller from a breadboard to a soldered perfboard PCB and adjusted constants accordingly. The MPU-6050 breakout board is now soldered directly to the PCB to keep it steady. The speed PID gains and max speed were altered to make driving smoother. Also updated the code when the robot falls over so that it can be stood up again and continue on.

Also moved the joystick remote from a breadboard to a soldered perfboard PCB, and adjusted the joystick orientation accordingly.